Create a utility class for static SnapshotsService methods
This moves the static methods out of the SnapshotsService to
make the stateful code more visible. This is one step towards
refactoring the SnapshotsService.

Relates ES-11650
